Move `load_file_as_dataframe` to shared module
Moves `UnsupportedFileExtensionError` to shared module as well since
it is raised by :func: `load_file_as_dataframe`
